Remove all MSVC signed-unsigned warnings with static_cast

+++ field.cpp
@@ -2331,7 +2331,7 @@ int32 field::check_tribute(card* pcard, int32 min, int32 max, group* mg, uint8 t
 	int32 ct = get_tofield_count(pcard, toplayer, LOCATION_MZONE, sumplayer, LOCATION_REASON_TOFIELD, zone);
 	if(ct <= 0 && max <= 0)
 		return FALSE;
-	for(auto& pcard : (ex_list.size() >= min) ? ex_list : release_list) {
+	for(auto& pcard : (static_cast<int>(ex_list.size()) >= min) ? ex_list : release_list) {
 		if(pcard->current.location == LOCATION_MZONE && pcard->current.controler == toplayer) {
 			s++;
 			if((zone >> pcard->current.sequence) & 1)
